+++ src/horizondata.tsx
@@ -1,0 +1,26 @@
+export const horizonData = [
+  {
+    imageUrl:
+      'https://assets.serlo.org/5c6ac35d6fe47_d00504b120098378a79473dfd0af2e93b7dbf605.png',
+    title: 'Öfters hier?',
+    text:
+      'Wenn serlo.org deine Lieblingslernplattform ist freuen wir uns von dir zu erfahren, wieso! Klicke hier um uns eine Nachricht zu hinterlassen. ',
+    url: 'https://de.serlo.org/horizon/37/out'
+  },
+  {
+    imageUrl:
+      'https://assets.serlo.org/5de5283a1c996_86b77d3d897f49ae20cd709d1bc51de97d15d42e.jpg',
+    title: 'Mitmachen bei Serlo',
+    text:
+      'Hinter serlo.org stehen viele engagierte Menschen, die Bildung besser und gerechter machen wollen. Auch du kannst mitmachen! Klicke hier, um zu erfahren, wie du Teil der Serlo Community werden kannst.',
+    url: 'https://de.serlo.org/horizon/24/out'
+  },
+  {
+    imageUrl:
+      'https://assets.serlo.org/5de528a9d79cd_7bbd720d6d5702451bd1c4ed760a603151035912.jpg',
+    title: 'Gib uns Feedback!',
+    text:
+      'Mit der Kommentar-Funktion kannst du uns zu jedem Inhalt sagen was dir gefällt - und was besser sein könnte. Du kannst auch Fragen stellen, wenn etwas unklar ist. Wir freuen uns über deinen Input! ',
+    url: 'https://de.serlo.org/horizon/23/out'
+  }
+]
